Renamed transactions view component.
