+++ src/client/App.jsx
@@ -2,12 +2,12 @@ import React from 'react';
 import {Containers} from '@opuscapita/service-base-ui';
 import {Route, Redirect} from 'react-router';
 import InvoiceArchive from './components/InvoiceArchive.jsx';
-import InvoiceTransaction from './components/InvoiceTransaction.jsx';
+import InvoiceArchiveDocument from './components/InvoiceArchiveDocument.jsx';
 
 const App = () =>
     <Containers.ServiceLayout serviceName="archive">
         <Route exact path='/invoices' component={InvoiceArchive}/>
-        <Route path='/invoices/:indexName/transactions/:id' component={InvoiceTransaction} />
+        <Route path='/invoices/:indexName/transactions/:id' component={InvoiceArchiveDocument} />
 
         <Redirect from='/' to='invoices' />
     </Containers.ServiceLayout>;
